Extract note path builder logic
