+++ src/services/note/NotePathBuilder.ts
@@ -1,0 +1,249 @@
+import { normalizePath } from "obsidian";
+import { BookMetadata, GroupingSettings } from "src/types";
+import { FileUtils } from "src/utils/FileUtils";
+import ObsidianHardcover from "src/main";
+
+export class NotePathBuilder {
+	constructor(
+		private fileUtils: FileUtils,
+		private plugin: ObsidianHardcover,
+	) {}
+
+	public generateNotePath(
+		bookMetadata: BookMetadata,
+		groupingSettings: GroupingSettings,
+		rawContributors?: Record<any, any>[],
+	): string {
+		const filename = this.fileUtils.processFilenameTemplate(
+			this.plugin.settings.filenameTemplate,
+			bookMetadata.variables,
+		);
+
+		let basePath = normalizePath(this.plugin.settings.targetFolder);
+
+		if (groupingSettings.enabled) {
+			const directories = this.buildDirectoryPath(
+				bookMetadata,
+				groupingSettings,
+				rawContributors,
+			);
+
+			if (directories) {
+				basePath = `${basePath}/${directories}`;
+			}
+		}
+
+		return basePath ? `${basePath}/${filename}` : filename;
+	}
+
+	public buildDirectoryPath(
+		bookMetadata: BookMetadata,
+		groupingSettings: GroupingSettings,
+		rawContributors?: Record<any, any>[],
+	): string {
+		const pathComponents: string[] = [];
+
+		if (
+			groupingSettings.groupBy === "author" ||
+			groupingSettings.groupBy === "author-series"
+		) {
+			const authorDirectory = this.getAuthorDirectory(
+				bookMetadata,
+				rawContributors,
+			);
+
+			if (authorDirectory) {
+				pathComponents.push(authorDirectory);
+			}
+		}
+
+		if (
+			groupingSettings.groupBy === "series" ||
+			groupingSettings.groupBy === "author-series"
+		) {
+			const seriesDirectory = this.getSeriesDirectory(bookMetadata);
+
+			if (seriesDirectory) {
+				pathComponents.push(seriesDirectory);
+			}
+		}
+
+		return pathComponents.join("/");
+	}
+
+	private getAuthorDirectory(
+		bookMetadata: BookMetadata,
+		rawContributors?: Record<any, any>[],
+	): string | null {
+		const authors = bookMetadata.variables.editionAuthors;
+
+		// check if multiple authors and should use collections folder
+		if (
+			Array.isArray(authors) &&
+			authors.length > 1 &&
+			this.plugin.settings.grouping.multipleAuthorsBehavior ===
+				"useCollectionsFolder"
+		) {
+			return this.fileUtils.sanitizeFolderName(
+				this.plugin.settings.grouping.collectionsFolderName,
+			);
+		}
+
+		if (Array.isArray(authors) && authors.length > 0) {
+			let authorName = authors[0].replace(/[\[\]']+/g, "");
+
+			// format the name based on settings
+			if (this.plugin.settings.grouping.authorFormat === "lastFirst") {
+				authorName = this.formatNameAsLastFirst(authorName);
+			}
+
+			return this.fileUtils.sanitizeFolderName(authorName);
+		}
+
+		//  if no authors and using fallback folder, return the folder name
+		if (
+			this.plugin.settings.grouping.noAuthorBehavior === "useFallbackFolder"
+		) {
+			return this.fileUtils.sanitizeFilename(
+				this.plugin.settings.grouping.fallbackFolderName,
+			);
+		}
+
+		// if no authors and using fallback priority, try to find Writer/Editor/first contributor
+		if (
+			this.plugin.settings.grouping.noAuthorBehavior ===
+				"useFallbackPriority" &&
+			rawContributors
+		) {
+			const fallbackName = this.findFallbackAuthor(rawContributors);
+			if (fallbackName) {
+				let authorName = fallbackName;
+
+				// format the name based on settings
+				if (this.plugin.settings.grouping.authorFormat === "lastFirst") {
+					authorName = this.formatNameAsLastFirst(authorName);
+				}
+
+				return this.fileUtils.sanitizeFolderName(authorName);
+			}
+		}
+
+		return null;
+	}
+
+	private getSeriesDirectory(bookMetadata: BookMetadata): string | null {
+		const series = bookMetadata.variables.series;
+
+		if (Array.isArray(series) && series.length > 0) {
+			let seriesName = series[0];
+
+			// extract series name from wikilink format: [[Series|Series #1]] -> Series
+			const wikilinkMatch = seriesName.match(
+				/^\[\[([^|\]]+)(?:\|[^\]]+)?\]\]$/,
+			);
+			if (wikilinkMatch) {
+				seriesName = wikilinkMatch[1];
+			} else {
+				// fallback: remove series position info if it exists ("Series Name #1" -> "Series Name")
+				seriesName = seriesName.replace(/\s*#\d+.*$/, "");
+			}
+
+			return this.fileUtils.sanitizeFolderName(seriesName);
+		}
+
+		return null;
+	}
+
+	private formatNameAsLastFirst(name: string): string {
+		name = name.trim();
+		// if name already contains a comma, assume it's already in "Last, First" format
+		if (name.includes(",")) {
+			return name;
+		}
+
+		const parts = name.split(/\s+/).filter((p) => p.length > 0);
+
+		if (parts.length === 1) {
+			return name;
+		}
+
+		// check for generational suffixes
+		const suffixes = [
+			"jr.",
+			"jr",
+			"junior",
+			"sr.",
+			"sr",
+			"senior",
+			"ii",
+			"iii",
+			"iv",
+			"v",
+			"vi",
+			"vii",
+			"viii",
+			"ix",
+			"x",
+		];
+
+		let lastNameIndex = parts.length - 1;
+		let suffix = "";
+
+		// check if last part is a suffix
+		if (suffixes.includes(parts[lastNameIndex].toLowerCase())) {
+			suffix = parts[lastNameIndex];
+			lastNameIndex--;
+
+			// safety check
+			if (lastNameIndex < 0) {
+				return name; // malformed, return as-is
+			}
+		}
+
+		const lastName = parts[lastNameIndex];
+		const firstName = parts.slice(0, lastNameIndex).join(" ");
+
+		return suffix
+			? `${lastName} ${suffix}, ${firstName}`
+			: `${lastName}, ${firstName}`;
+	}
+
+	private findFallbackAuthor(
+		contributorsData: Record<any, any>[],
+	): string | null {
+		if (
+			!contributorsData ||
+			!Array.isArray(contributorsData) ||
+			contributorsData.length === 0
+		) {
+			return null;
+		}
+
+		// try Writer
+		const writers = contributorsData
+			.filter((item) => item.contribution === "Writer")
+			.map((item) => item.author?.name)
+			.filter((name) => !!name);
+
+		if (writers.length > 0) {
+			return writers[0];
+		}
+
+		// try Editor
+		const editors = contributorsData
+			.filter((item) => item.contribution === "Editor")
+			.map((item) => item.author?.name)
+			.filter((name) => !!name);
+
+		if (editors.length > 0) {
+			return editors[0];
+		}
+
+		// use first contributor available
+		const firstContributor = contributorsData
+			.map((item) => item.author?.name)
+			.filter((name) => !!name)[0];
+
+		return firstContributor || null;
+	}
+}
